add c param type to conversions and fix bugs
